Update UI structure.
