+++ packages/app/src/wrappers/page.tsx
@@ -1,0 +1,11 @@
+import { PageProps, PluginOptions } from 'gatsby';
+import React from 'react';
+
+type Props = {
+  pluginOptions: PluginOptions;
+} & PageProps;
+const PageWrapper: React.FC<Props> = ({ children }) => {
+  return <>{children}</>;
+};
+
+export default PageWrapper;
